+++ test/hexapdf/type/acro_form/test_form.rb
@@ -159,11 +159,15 @@ describe HexaPDF::Type::AcroForm::Form do
     end
 
     def applies_variable_text_properties(method, **args)
+      field = @acro_form.send(method, "field", **args)
+      font_name, font_size, font_color = field.parse_default_appearance_string
+      assert_equal(:'Helvetica', @acro_form.default_resources.font(font_name)[:BaseFont])
+      assert_equal(0, font_size)
+      assert_equal(HexaPDF::Content::ColorSpace::DeviceGray.new.color(0), font_color)
+
       field = @acro_form.send(method, "field", **args, font: 'Times')
       font_name, font_size, font_color = field.parse_default_appearance_string
       assert_equal(:'Times-Roman', @acro_form.default_resources.font(font_name)[:BaseFont])
-      assert_equal(0, font_size)
-      assert_equal(HexaPDF::Content::ColorSpace::DeviceGray.new.color(0), font_color)
 
       field = @acro_form.send(method, "field", **args, font_options: {variant: :bold})
       font_name, = field.parse_default_appearance_string
@@ -171,7 +175,6 @@ describe HexaPDF::Type::AcroForm::Form do
 
       field = @acro_form.send(method, "field", **args, font_size: 10)
       font_name, font_size = field.parse_default_appearance_string
-      assert_equal(:Helvetica, @acro_form.default_resources.font(font_name)[:BaseFont])
       assert_equal(10, font_size)
 
       field = @acro_form.send(method, "field", **args, font_color: "red")